requirements.txt
pycryptodome>=3.19
zstandard>=0.22
